requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.53
psycopg2-binary==2.9.10
pandas==2.3.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
openpyxl==3.1.5
xlsxwriter==3.2.0
pydantic[email]==2.7.1
